Fix "Launch" action for updated Homebrew JSON format
The format of `artifacts` has changed.
I can't find where, when or if this was documented (release notes?), but it's here now.
(The new format does make marginally more sense? I think)

This code is still just a basic heuristic that doesn't do that great a job.
e.g., it misses apps installed via `pkg`s